server/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Markdown==3.7
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
uritemplate==4.1.1
